dom/base/test/test_bug711047.html

<!DOCTYPE html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=711047
-->
<title>Test for Bug 711047</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=711047">Mozilla Bug 711047</a>
<div id="content">
</div>
<pre id="test">
<script>
/** Test for Bug 711047 **/
ok(!("RangeException" in window), "RangeException shouldn't exist");
</script>
</pre>
